refactor: package naming convention 반영
