+++ perfume-api/src/main/java/io/perfume/api/common/config/GeneratorConfiguration.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.configurations;
+package io.perfume.api.common.config;
 
 import generator.Generator;
 import generator.impl.NumberGenerator;
