+++ perfume-api/src/main/java/io/perfume/api/common/config/JwtAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.configurations;
+package io.perfume.api.common.config;
 
 import jwt.JsonWebTokenGenerator;
 import org.jetbrains.annotations.NotNull;
