+++ perfume-api/src/main/java/io/perfume/api/common/config/MailSenderConfiguration.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.configurations;
+package io.perfume.api.common.config;
 
 import mailer.MailSender;
 import mailer.impl.SESMailSender;
