+++ perfume-api/src/main/java/io/perfume/api/common/config/WhiteListConfiguration.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.configurations;
+package io.perfume.api.common.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
